+++ src/test/java/com/amazin/amazinonlinebookstore/BookControllerTest.java
@@ -14,12 +14,15 @@ import static org.mockito.Mockito.*;
 
 public class BookControllerTest {
 
+    // Mocking the BookRepository to simulate database interactions
     @Mock
     private BookRepository bookRepository;
 
+    // Injecting the mocked BookRepository into the BookController
     @InjectMocks
     private  BookController bookController;
 
+    // Initialize mocks before each test
     @BeforeEach
     void setUp() {
         MockitoAnnotations.openMocks(this);
@@ -27,22 +30,28 @@ public class BookControllerTest {
 
     @Test
     public void testAddBook() {
+        // Create a book object and configure the mock repository behavior
         Book book = new Book("Effective Java", "Best practices for Java", "Joshua Bloch", "8647823308954", LocalDate.of(2024, 1, 1), 42.99);
         when(bookRepository.save(book)).thenReturn(book);
 
+        // Call the addBook method in the controller
         Book savedBook = bookController.addBook(book);
 
+        // Verify the returned book is not noll and has the expected title
         assertNotNull((savedBook));
         assertEquals("Effective Java", savedBook.getTitle());
     }
 
     @Test
     public void testRemoveBook(){
+        // Create a book object and configure the mock repository behavior
         Book book = new Book("Effective Java", "Best practices for Java", "Joshua Bloch", "8647823308954", LocalDate.of(2024, 1, 1), 42.99);
         when(bookRepository.findById(1L)).thenReturn(Optional.of(book));
 
-        bookController.removeBook(1L);
+        // Call the removeBook method in the controller
+        bookController.removeBook(1L); // 1L = ID 1, long
 
+        // Verify the delete method on the repository is called exactly once
         verify(bookRepository, times(1)).delete(book);
     }
