+++ src/test/java/com/amazin/amazinonlinebookstore/BookTest.java
@@ -10,9 +10,25 @@ import static org.junit.jupiter.api.Assertions.*;
 public class BookTest {
 
     @Test
+    public void testDefaultConstructor(){
+        // Create new book with default attributes
+        Book book = new Book();
+
+        // Check that the book attributes are correct (everything is null, except price which is 0)
+        assertNull(book.getTitle());
+        assertNull(book.getDescription());
+        assertNull(book.getAuthor());
+        assertNull(book.getISBN());
+        assertNull(book.getPublishDate());
+        assertEquals(0, book.getPrice());
+    }
+
+    @Test
     public void testBookCreation(){
+        // Create new book with custom attributes
         Book book = new Book("Effective Java",  "Best practices for Java", "Joshua Bloch","8647823308954", LocalDate.of(2024, 1, 1), 42.99);
 
+        // Check that the book has the correct attributes
         assertEquals("Effective Java", book.getTitle());
         assertEquals("Best practices for Java", book.getDescription());
         assertEquals("Joshua Bloch", book.getAuthor());
@@ -23,8 +39,10 @@ public class BookTest {
 
     @Test
     public void testBookSettersAndGetters() {
+        // Create new book
         Book book = new Book();
 
+        // Set various attributes using setters
         book.setTitle("Effective Java");
         book.setDescription("Best practices for Java");
         book.setAuthor("Joshua Bloch");
@@ -32,6 +50,7 @@ public class BookTest {
         book.setPrice(42.99);
         book.setPublishDate(LocalDate.of(2024,1,1));
 
+        // Verify that the getters return the correct values
         assertEquals("Effective Java", book.getTitle());
         assertEquals("Best practices for Java", book.getDescription());
         assertEquals("Joshua Bloch", book.getAuthor());
@@ -42,8 +61,10 @@ public class BookTest {
 
     @Test
     public void testBookToString(){
+        // Create new book
         Book book = new Book("Effective Java",  "Best practices for Java", "Joshua Bloch","8647823308954", LocalDate.of(2024, 1, 1), 42.99);
 
+        // Verify that the toString method matches the expected string format
         String expectedString = "\nTitle: Effective Java\nAuthor: Joshua Bloch\nPublish Date: 2024-01-01\nISBN: 8647823308954\nPrice: 42.99";
         assertEquals(expectedString, book.toString());
     }
